requirements.txt
annotated-types==0.7.0
anyio==4.12.0
bcrypt==4.0.1
click==8.3.1
ecdsa==0.19.1
fastapi==0.110.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
passlib==1.7.4
psycopg-binary==3.2.13
pyasn1==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.29
starlette==0.36.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.29.0
